// apps/web/src/api/generated/organizer-tournaments/createGamingRoomDtoPurpose.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type CreateGamingRoomDtoPurpose = typeof CreateGamingRoomDtoPurpose[keyof typeof CreateGamingRoomDtoPurpose];


export const CreateGamingRoomDtoPurpose = {
  COMPETITION: 'COMPETITION',
  PRACTICE: 'PRACTICE',
  WARMUP: 'WARMUP',
  FINAL_STAGE: 'FINAL_STAGE',
  BACKUP: 'BACKUP',
} as const;