// apps/web/src/api/generated/organizer-tournaments/createGamingRoomDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { CreateGamingRoomDtoPurpose } from './createGamingRoomDtoPurpose';

export interface CreateGamingRoomDto {
  /**
     * @minLength 2
     * @maxLength 150
     */
  name: string;
  /** @maxLength 1000 */
  description?: string;
  purpose: CreateGamingRoomDtoPurpose;
  /**
     * @minimum 1
     * @maximum 1000
     */
  stationCount: number;
  /**
     * @minLength 2
     * @maxLength 150
     */
  cpu: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  gpu: string;
  /** @maxLength 100 */
  ram?: string;
  /** @maxLength 100 */
  storage?: string;
  /** @maxLength 100 */
  operatingSystem?: string;
  /** @maxLength 100 */
  monitorBrand?: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  monitorModel: string;
  /**
     * @minimum 1
     * @maximum 100
     */
  monitorSizeInches?: number;
  /** @maxLength 50 */
  monitorResolution?: string;
  /**
     * @minimum 30
     * @maximum 1000
     */
  monitorRefreshRateHz: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  monitorResponseTimeMs?: number;
  /**
     * @minLength 2
     * @maxLength 150
     */
  mouse: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  keyboard: string;
  /**
     * @minLength 2
     * @maxLength 150
     */
  headset: string;
  /** @maxLength 150 */
  mousePad?: string;
  /** @maxLength 150 */
  controller?: string;
  /** @maxLength 300 */
  internetConnection?: string;
  /** @maxLength 2000 */
  equipmentNotes?: string;
}