// apps/web/src/api/generated/organizer-tournaments/gamingRoomResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { GamingRoomMonitorDetailsDto } from './gamingRoomMonitorDetailsDto';
import type { GamingRoomPcSpecsDto } from './gamingRoomPcSpecsDto';
import type { GamingRoomPeripheralDetailsDto } from './gamingRoomPeripheralDetailsDto';
import type { GamingRoomResponseDtoPurpose } from './gamingRoomResponseDtoPurpose';

export interface GamingRoomResponseDto {
  id: string;
  venueId: string;
  name: string;
  /** @nullable */
  description?: string | null;
  purpose: GamingRoomResponseDtoPurpose;
  stationCount: number;
  pcSpecs: GamingRoomPcSpecsDto;
  monitor: GamingRoomMonitorDetailsDto;
  peripherals: GamingRoomPeripheralDetailsDto;
  /** @nullable */
  internetConnection?: string | null;
  /** @nullable */
  equipmentNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}